refactor: config editor
