+++ dashboard/src/views/site/SiteConfig.vue
@@ -35,79 +35,23 @@
 					Save changes
 				</Button>
 			</template>
-			<div class="flex space-x-4">
-				<div class="w-full shrink-0 space-y-4 md:w-2/3">
-					<div class="space-y-4" v-if="editMode">
-						<div
-							class="grid grid-cols-5 gap-4"
-							v-for="(config, i) in $resources.siteConfig.data"
-							:key="i"
-						>
-							<Input
-								type="text"
-								placeholder="key"
-								v-model="config.key"
-								class="col-span-2"
-								@change="isDirty = true"
-							/>
-							<Input
-								type="select"
-								placeholder="type"
-								v-model="config.type"
-								:options="['String', 'Number', 'JSON', 'Boolean']"
-								@change="onTypeChange(config)"
-							/>
-							<div class="col-span-2 flex items-center">
-								<Input
-									class="w-full"
-									v-bind="configInputProps(config)"
-									:input-class="{ 'font-mono': config.type === 'JSON' }"
-									placeholder="value"
-									v-model="config.value"
-									@change="isDirty = true"
-								/>
-								<button
-									class="ml-2 rounded-md p-1 hover:bg-gray-100"
-									@click="removeConfig(config)"
-								>
-									<FeatherIcon name="x" class="h-5 w-5 text-gray-700" />
-								</button>
-							</div>
-						</div>
-						<ErrorMessage :message="$resources.updateSiteConfig.error" />
-						<div class="space-x-2">
-							<Button @click="addConfig" v-if="!isDirty"> Add Key </Button>
-						</div>
-					</div>
-					<div v-else>
-						<Form
-							v-if="readOnlyFormProps.fields && readOnlyFormProps.fields.length"
-							v-bind="readOnlyFormProps"
-							class="pointer-events-none"
-						/>
-						<span class="text-base text-gray-600" v-else>
-							No keys added. Click on Edit Config to add one.
-						</span>
-					</div>
-				</div>
-				<div
-					class="hidden max-w-full flex-1 overflow-x-scroll whitespace-pre-line rounded bg-gray-100 p-4 font-mono text-base md:block"
-				>
-					<div class="mb-4">site_config.json</div>
-					<div v-html="siteConfigPreview"></div>
-				</div>
-			</div>
+			<ConfigEditor
+				:configData="$resources.siteConfig.data"
+				:standardConfigKeys="$resources.standardConfigKeys.data"
+				:editMode="editMode"
+				@isDirty="val => (isDirty = val)"
+			/>
 		</Card>
 	</div>
 </template>
 
 <script>
-import Form from '@/components/Form.vue';
+import ConfigEditor from '@/components/ConfigEditor.vue';
 
 export default {
 	name: 'SiteConfig',
 	components: {
-		Form
+		ConfigEditor
 	},
 	props: ['site'],
 	data() {
@@ -184,41 +128,6 @@ export default {
 		}
 	},
 	methods: {
-		configInputProps(config) {
-			let type = {
-				String: 'text',
-				Number: 'number',
-				JSON: 'textarea',
-				Boolean: 'select'
-			}[config.type];
-			return {
-				type,
-				options: config.type === 'Boolean' ? ['1', '0'] : null
-			};
-		},
-		addConfig() {
-			this.$resources.siteConfig.data.push({
-				key: '',
-				value: '',
-				type: 'String'
-			});
-			this.isDirty = true;
-		},
-		removeConfig(config) {
-			this.$resources.siteConfig.data = this.$resources.siteConfig.data.filter(
-				d => d !== config
-			);
-			this.isDirty = true;
-		},
-		onTypeChange(config) {
-			if (config.type === 'Boolean') {
-				config.value = '1';
-			} else if (config.type === 'Number') {
-				config.value = Number(config.value) || 0;
-			} else if (config.type === 'String') {
-				config.value = String(config.value);
-			}
-		},
 		updateSiteConfig() {
 			if (this.isDirty) {
 				this.$resources.updateSiteConfig.submit();
@@ -226,58 +135,6 @@ export default {
 				this.editMode = false;
 				this.isDirty = false;
 			}
-		}
-	},
-	computed: {
-		siteConfigPreview() {
-			let obj = {};
-
-			for (let d of this.$resources.siteConfig.data) {
-				let value = d.value;
-				if (['Boolean', 'Number'].includes(d.type)) {
-					value = Number(d.value);
-				} else if (d.type === 'JSON') {
-					try {
-						value = JSON.parse(d.value);
-					} catch (error) {
-						value = {};
-					}
-				}
-				obj[d.key] = value;
-			}
-			return JSON.stringify(obj, null, '&nbsp; ');
-		},
-		readOnlyFormProps() {
-			if (!this.$resources.standardConfigKeys.data) {
-				return {};
-			}
-
-			let fields = this.$resources.siteConfig.data.map(config => {
-				let standardKey = this.$resources.standardConfigKeys.data.find(
-					d => d.key === config.key
-				);
-				return {
-					label: standardKey?.title || config.key,
-					fieldname: standardKey?.key || config.key
-				};
-			});
-
-			let modelValue = {};
-			for (let d of this.$resources.siteConfig.data) {
-				let value = d.value;
-				if (['Boolean', 'Number'].includes(d.type)) {
-					value = Number(value);
-				}
-				modelValue[d.key] = value;
-			}
-
-			return {
-				fields,
-				modelValue
-			};
-		},
-		NotAllowed() {
-			return `Not Permitted in ${this.site.status} mode`;
 		}
 	}
 };
